swap the endpoint assignment
assignment of streamline begin to cortical target0 and streamline end to cortical target1 did not result in valid streamlines in our data for ATR. Swapping did. 
This has been briefly tested for a few other tracts, but I am far from sure that this change is valid for all. In fact, I believe that the assignment should be tested routinely as part of the debug-mode: distances between zero end and both cortical target and distances of -1 end and both targets are recorded and if the distribution of distances is clearly different, and does not follow the expected order , the user should be notified to arrange the targets (or rois) in a different sequence.