Use Tableau 20 colors for the 20 waypoint-defined bundles.
It turns out that we don't need 40 colors. This is only needed in AFQ-Browser
because we need to be able to highlight the picked bundles.

This means that the code can be simplified quite a bit. This commit also
clarifies and simplifies the logic, including comments. And adds a docstring.

Changes in the API module to remain consistent.