BF: _gen_sl_counts was not able to deal with cases with more than one subject.
Also, was generating sl_counts of 0 in my experiments with it.

I think that the failure with more than one subject arises because
when one subject has that added to their row, it gets added to the columns
for the entire dataframe, but no value gets added in other rows. So,
when pandas gets to the next row it freaks out.

I suspect this is also the reason that the nightly3 tests are currently failing.

At any rate, I think this fix is more consistent with how we are doing
things the rest of the time. The only challenge is that it requires one
more read of the entire segmented tractogram from file.